feat: add light constructor
Adds a `createHeliaLight` factory function that creates a minimally
configured helia node.

No routing, only supports dag-pb and raw codecs, only sha256 and
identity hashes.

Essentially the minimum requirements for consuming IPFS content.
