+++ packages/helia/package.json
@@ -49,13 +49,46 @@
     "prepublishOnly": "node scripts/update-version.js && npm run build"
   },
   "dependencies": {
+    "@helia/bitswap": "^3.2.3",
+    "@helia/http": "^3.1.4",
     "@helia/interface": "^6.2.1",
+    "@helia/libp2p": "^0.0.0",
     "@helia/utils": "^2.5.2",
-    "multiformats": "^14.0.0"
+    "@ipld/dag-cbor": "^10.0.1",
+    "@ipld/dag-json": "^11.0.0",
+    "@ipld/dag-pb": "^4.1.7",
+    "@ipshipyard/crypto": "^1.1.0",
+    "@ipshipyard/keychain": "^1.0.2",
+    "@libp2p/interface": "^3.2.4",
+    "@libp2p/utils": "^7.2.3",
+    "@multiformats/dns": "^1.0.15",
+    "@multiformats/multiaddr": "^13.0.3",
+    "abort-error": "^1.0.2",
+    "any-signal": "^4.2.0",
+    "birnam": "^1.0.0",
+    "blockstore-core": "^7.0.1",
+    "cborg": "^5.1.4",
+    "datastore-core": "^12.0.1",
+    "interface-blockstore": "^7.0.1",
+    "interface-datastore": "^10.0.1",
+    "it-drain": "^3.0.12",
+    "it-filter": "^3.1.6",
+    "it-foreach": "^2.1.7",
+    "it-merge": "^3.0.14",
+    "main-event": "^1.0.4",
+    "mortice": "^3.3.1",
+    "multiformats": "^14.0.0",
+    "progress-events": "^1.1.0",
+    "uint8arrays": "^6.1.1"
   },
   "devDependencies": {
     "aegir": "^48.0.11",
-    "interface-datastore": "^10.0.1"
+    "delay": "^7.0.0",
+    "it-all": "^3.0.11",
+    "it-map": "^3.1.6",
+    "it-to-buffer": "^5.0.0",
+    "sinon": "^22.0.0",
+    "sinon-ts": "^2.0.0"
   },
   "sideEffects": false
 }
