+++ packages/helia/src/pins.ts
@@ -1,3 +1,4 @@
+import { AlreadyPinnedError, depthFirstWalker } from '@helia/utils'
 import { InvalidParametersError } from '@libp2p/interface'
 import * as cborg from 'cborg'
 import { Key } from 'interface-datastore'
@@ -5,8 +6,6 @@ import { base36 } from 'multiformats/bases/base36'
 import { CID } from 'multiformats/cid'
 import { CustomProgressEvent } from 'progress-events'
 import { equals as uint8ArrayEquals } from 'uint8arrays/equals'
-import { AlreadyPinnedError } from './errors.ts'
-import { depthFirstWalker } from './graph-walker.ts'
 import type { CodecLoader } from '@helia/interface'
 import type { AddOptions, IsPinnedOptions, LsOptions, Pin, Pins, RmOptions } from '@helia/interface'
 import type { AbortOptions } from '@libp2p/interface'
