+++ packages/helia/src/routing.ts
@@ -1,10 +1,10 @@
 import { NoRoutersAvailableError } from '@helia/interface'
+import { GetFailedError } from '@helia/utils'
 import { NotFoundError, start, stop } from '@libp2p/interface'
 import { Queue } from '@libp2p/utils'
 import merge from 'it-merge'
 import { CustomProgressEvent } from 'progress-events'
 import { toString as uint8ArrayToString } from 'uint8arrays/to-string'
-import { GetFailedError } from './errors.ts'
 import type { Routing as RoutingInterface, Provider, RoutingOptions, RoutingFindProvidersProgressEvents, RoutingProvideProgressEvents, RoutingPutProgressEvents, RoutingGetProgressEvents, RoutingFindPeerProgressEvents, RoutingGetClosestPeersProgressEvents, RoutingCancelReprovideProgressEvents, Router, Peer } from '@helia/interface'
 import type { ComponentLogger, Logger, Metrics, Startable } from '@libp2p/interface'
 import type { AbortOptions } from 'abort-error'
