+++ packages/helia/test/block-broker.spec.ts
@@ -1,6 +1,6 @@
 import { stop } from '@libp2p/interface'
-import { defaultLogger } from '@libp2p/logger'
 import { expect } from 'aegir/chai'
+import { defaultLogger } from 'birnam'
 import { MemoryBlockstore } from 'blockstore-core'
 import delay from 'delay'
 import all from 'it-all'
@@ -9,8 +9,8 @@ import toBuffer from 'it-to-buffer'
 import * as raw from 'multiformats/codecs/raw'
 import Sinon from 'sinon'
 import { stubInterface } from 'sinon-ts'
-import { getHasher } from '../src/utils/get-hasher.ts'
-import { NetworkedStorage } from '../src/utils/networked-storage.ts'
+import { getHasher } from '../src/get-hasher.ts'
+import { NetworkedStorage } from '../src/networked-storage.ts'
 import { createBlock } from './fixtures/create-block.ts'
 import type { BlockBroker } from '@helia/interface'
 import type { Blockstore } from 'interface-blockstore'
