+++ packages/helia/test/networked-storage.spec.ts
@@ -1,5 +1,5 @@
-import { defaultLogger } from '@libp2p/logger'
 import { expect } from 'aegir/chai'
+import { defaultLogger } from 'birnam'
 import { MemoryBlockstore } from 'blockstore-core'
 import delay from 'delay'
 import all from 'it-all'
@@ -14,11 +14,11 @@ import Sinon from 'sinon'
 import { stubInterface } from 'sinon-ts'
 import { fromString as uint8ArrayFromString } from 'uint8arrays/from-string'
 import { toString as uint8ArrayToString } from 'uint8arrays/to-string'
-import { getHasher } from '../../src/utils/get-hasher.ts'
-import { NetworkedStorage } from '../../src/utils/networked-storage.ts'
-import { getCidBlockVerifierFunction } from '../../src/utils/storage.ts'
-import { createBlock } from '../fixtures/create-block.ts'
-import type { NetworkedStorageComponents } from '../../src/utils/networked-storage.ts'
+import { getHasher } from '../src/get-hasher.ts'
+import { NetworkedStorage } from '../src/networked-storage.ts'
+import { getCidBlockVerifierFunction } from '../src/storage.ts'
+import { createBlock } from './fixtures/create-block.ts'
+import type { NetworkedStorageComponents } from '../src/networked-storage.ts'
 import type { BlockBroker } from '@helia/interface'
 import type { Blockstore } from 'interface-blockstore'
 import type { StubbedInstance } from 'sinon-ts'
