+++ packages/helia/test/storage.spec.ts
@@ -8,9 +8,9 @@ import map from 'it-map'
 import toBuffer from 'it-to-buffer'
 import * as raw from 'multiformats/codecs/raw'
 import { stubInterface } from 'sinon-ts'
+import { BlockStorage } from '../src/block-storage.ts'
+import { getCodec } from '../src/get-codec.ts'
 import { PinsImpl } from '../src/pins.ts'
-import { BlockStorage } from '../src/storage.ts'
-import { getCodec } from '../src/utils/get-codec.ts'
 import { createBlock } from './fixtures/create-block.ts'
 import type { Blocks, Routing, SessionBlockstore } from '@helia/interface'
 import type { Pins } from '@helia/interface'
