+++ packages/interop/package.json
@@ -66,8 +66,10 @@
     "@helia/unixfs": "^7.2.1",
     "@ipld/car": "^5.4.6",
     "@ipld/dag-cbor": "^10.0.1",
+    "@ipld/dag-json": "^11.0.0",
     "@ipld/dag-pb": "^4.1.7",
     "@libp2p/floodsub": "^11.0.23",
+    "@libp2p/identify": "^4.1.8",
     "@libp2p/interface": "^3.2.3",
     "@libp2p/kad-dht": "^16.3.2",
     "@libp2p/keychain": "^6.1.2",
