+++ packages/libp2p/package.json
@@ -80,10 +80,10 @@
   },
   "devDependencies": {
     "aegir": "^48.0.11",
-    "helia": "^6.1.4",
+    "birnam": "^1.0.0",
+    "datastore-core": "^12.0.1",
     "it-drain": "^3.0.12",
-    "sinon-ts": "^2.0.0",
-    "uint8arrays": "^6.1.1"
+    "sinon-ts": "^2.0.0"
   },
   "browser": {
     "./dist/src/utils/libp2p-defaults.js": "./dist/src/utils/libp2p-defaults.browser.js"
