+++ packages/utils/package.json
@@ -49,46 +49,29 @@
   },
   "dependencies": {
     "@helia/interface": "^6.2.1",
-    "@ipld/dag-cbor": "^10.0.1",
-    "@ipld/dag-json": "^11.0.0",
-    "@ipld/dag-pb": "^4.1.7",
-    "@ipshipyard/crypto": "^1.1.0",
-    "@ipshipyard/keychain": "^1.0.2",
     "@libp2p/interface": "^3.2.3",
     "@libp2p/utils": "^7.2.2",
-    "@multiformats/dns": "^1.0.13",
     "@multiformats/multiaddr": "^13.0.3",
-    "abort-error": "^1.0.2",
-    "any-signal": "^4.2.0",
-    "birnam": "^1.0.0",
-    "blockstore-core": "^7.0.1",
-    "cborg": "^5.1.1",
-    "datastore-core": "^12.0.1",
-    "interface-blockstore": "^7.0.1",
-    "interface-datastore": "^10.0.1",
-    "it-drain": "^3.0.12",
     "it-filter": "^3.1.6",
-    "it-foreach": "^2.1.7",
-    "it-merge": "^3.0.14",
     "it-to-buffer": "^5.0.0",
-    "mortice": "^3.3.1",
     "multiformats": "^14.0.0",
     "p-defer": "^4.0.1",
     "progress-events": "^1.1.0",
-    "race-signal": "^2.0.0",
-    "uint8arrays": "^6.1.1"
+    "race-signal": "^2.0.0"
   },
   "devDependencies": {
+    "@ipld/dag-cbor": "^10.0.1",
     "@libp2p/crypto": "^5.1.19",
     "@libp2p/logger": "^6.2.8",
     "@libp2p/peer-id": "^6.0.10",
     "@types/sinon": "^21.0.1",
     "aegir": "^48.0.11",
+    "blockstore-core": "^7.0.1",
     "delay": "^7.0.0",
+    "interface-blockstore": "^7.0.1",
     "it-all": "^3.0.11",
     "it-map": "^3.1.6",
-    "sinon": "^22.0.0",
-    "sinon-ts": "^2.0.0"
+    "sinon": "^22.0.0"
   },
   "sideEffects": false
 }
