+++ packages/utils/src/is-promise.ts
