--- packages/utils/src/utils/constants.ts
@@ -1,3 +1,0 @@
-export const SALT_LENGTH = 16
-export const KEY_SIZE = 32
-export const ITERATIONS = 10000
